Add init_starknet_bootstrap helper to simplify tests
