+++ relayer/crates/starknet-cli/src/impls/bootstrap/starknet_chain.rs
@@ -20,6 +20,8 @@ pub struct BootstrapStarknetChainArgs {
     pub chain_command_path: String,
 
     pub erc20_contract_path: String,
+
+    pub ics20_contract_path: String,
 }
 
 pub struct LoadStarknetBootstrap;
@@ -47,11 +49,21 @@ where
             serde_json::from_str(&contract_str).map_err(App::raise_error)?
         };
 
+        let ics20_contract = {
+            let contract_str = runtime
+                .read_file_as_string(&args.ics20_contract_path.clone().into())
+                .await
+                .map_err(App::raise_error)?;
+
+            serde_json::from_str(&contract_str).map_err(App::raise_error)?
+        };
+
         let bootstrap = StarknetBootstrap {
             runtime: runtime.clone(),
             chain_command_path: args.chain_command_path.clone().into(),
             chain_store_dir: args.chain_store_dir.clone().into(),
             erc20_contract,
+            ics20_contract,
         };
 
         Ok(bootstrap)
