+++ relayer/crates/starknet-integration-tests/src/contexts/bootstrap.rs
@@ -66,6 +66,7 @@ pub struct StarknetBootstrap {
     pub chain_command_path: PathBuf,
     pub chain_store_dir: PathBuf,
     pub erc20_contract: SierraClass,
+    pub ics20_contract: SierraClass,
 }
 
 delegate_components! {
