+++ relayer/crates/starknet-integration-tests/src/lib.rs
@@ -1,6 +1,7 @@
 #![recursion_limit = "256"]
 
 pub mod contexts;
+pub mod utils;
 
 #[cfg(test)]
 pub mod tests;
