+++ relayer/crates/starknet-integration-tests/src/utils/mod.rs
@@ -1,0 +1,3 @@
+mod init_bootstrap;
+
+pub use init_bootstrap::*;
